app/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
Pillow==10.0.0
di4injector==0.0.1
slf4py==0.0.2
requests==2.31.0
boto3==1.28.49